Add files via upload
This is a script game using python. It's a simple project for beginners.
It's a game that the player throw a dice. If he/she got any number but 1, it will be added to his/her total score. But if he/she got number 1, his/her total score will
restart at zero and his/her round will finish. Now it's the other player turn.
If the player didn't want to throw the dice and save his/her total score, his/her turn will finish and the other player will start.
The game will finish when one of the two players has collected 25 or above
in his/her total score.
If the player entered a wrong input, the game will give a caution and then it will continue.